+++ DiFfRG/include/DiFfRG/physics/interpolation/linear_interpolator_3D.hh
@@ -73,9 +73,9 @@ namespace DiFfRG
             "LinearInterpolator3D: You probably called update() on a copied instance. This is not allowed. "
             "You need to call update() on the original instance.");
       // Populate host mirror
-      for (uint i = 0; i < sizes[0]; ++i)
-        for (uint j = 0; j < sizes[1]; ++j)
-          for (uint k = 0; k < sizes[2]; ++k)
+      for (size_t i = 0; i < sizes[0]; ++i)
+        for (size_t j = 0; j < sizes[1]; ++j)
+          for (size_t k = 0; k < sizes[2]; ++k)
             host_data(i, j, k) = static_cast<NT>(in_data[i * sizes[1] * sizes[2] + j * sizes[2] + k]);
       // Copy data to device
       Kokkos::deep_copy(device_data, host_data);
@@ -106,11 +106,11 @@ namespace DiFfRG
       idx_z = max(static_cast<decltype(idx_z)>(0), min(idx_z, static_cast<decltype(idx_z)>(sizes[2] - 1)));
 
       // Clamp the (upper) index to the range [1, sizes - 1]
-      uint x1 = static_cast<uint>(
+      size_t x1 = static_cast<size_t>(
           min(ceil(idx_x + static_cast<decltype(idx_x)>(1e-16)), static_cast<decltype(idx_x)>(sizes[0] - 1)));
-      uint y1 = static_cast<uint>(
+      size_t y1 = static_cast<size_t>(
           min(ceil(idx_y + static_cast<decltype(idx_y)>(1e-16)), static_cast<decltype(idx_y)>(sizes[1] - 1)));
-      uint z1 = static_cast<uint>(
+      size_t z1 = static_cast<size_t>(
           min(ceil(idx_z + static_cast<decltype(idx_z)>(1e-16)), static_cast<decltype(idx_z)>(sizes[2] - 1)));
 
       const auto corner000 = device_data(x1 - 1, y1 - 1, z1 - 1);
@@ -181,8 +181,8 @@ namespace DiFfRG
 
   private:
     const Coordinates coordinates;
-    const device::array<uint, 3> sizes;
-    const uint total_size;
+    const device::array<size_t, 3> sizes;
+    const size_t total_size;
 
     using ViewType = Kokkos::View<NT ***, DefaultMemorySpace, Kokkos::MemoryTraits<Kokkos::RandomAccess>>;
     using HostViewType = typename ViewType::HostMirror;
